Fix: Use the resource group specified by user for all resources (#27)
The copier template requested the user to specify a resource group for
the storage account, however, a random resource group would be created
and used for other resources, such as the kubernetes cluster. The option
specified by the user is now used everywhere where a resource group is
required.

Since the resource group is now assumed to be pre-existing, the
declaration in the Terraform files has to be changed from `resource` to
`data`, because if defined as a `resource` it signals that Terraform is
supposed to claim ownership of it.