Increase user server startup time and CPU resources
Currently the startup of a user server often times out when the
kubernetes service takes longer to assign the CPU resources. Increasing
the startup timeout to 10 mins is still acceptable and leads to fewer
failures.

We also increase the number of guaranteed and the upper limit of CPUs,
to make sure each user is able to run a workflow in a reasonable time.